+++ packages/sdks/e2e/angular-ssr/src/app/app.module.ts
@@ -3,8 +3,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { Blocks, Content } from '@builder.io/sdk-angular';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app.routing';
-import { BuilderBlockWithClassNameComponent } from './builder-block-with-class-name.component';
 import { CatchAllComponent } from './catch-all.component';
+import { BuilderBlockWithClassNameComponent } from './custom-components/builder-block-with-class-name.component';
 
 @NgModule({
   declarations: [
