+++ packages/sdks/e2e/angular-ssr/src/app/custom-components/builder-block-with-class-name.component.ts
