+++ packages/sdks/e2e/angular-ssr/src/app/custom-components/hello.component.ts
