+++ packages/sdks/e2e/angular/src/app/app.component.ts
@@ -6,8 +6,7 @@ import {
   type RegisteredComponent,
 } from '@builder.io/sdk-angular';
 import { getProps } from '@sdk/tests';
-import { BuilderBlockWithClassNameComponent } from './builder-block-with-class-name.component';
-import { HelloComponent } from './hello.component';
+import { customComponents } from './custom-components';
 
 interface BuilderProps {
   apiVersion: string;
@@ -49,66 +48,7 @@ export class AppComponent {
   content: BuilderProps['content'];
   data: BuilderProps['data'];
 
-  customComponents: RegisteredComponent[] = [
-    {
-      component: HelloComponent,
-      name: 'Hello',
-      inputs: [],
-      defaultChildren: [
-        {
-          '@type': '@builder.io/sdk:Element',
-          '@version': 2,
-          id: 'builder-ebca7d55d34f4fc9a6536600959cef5d',
-          component: {
-            name: 'Text',
-            options: {
-              text: 'inside an h1',
-            },
-          },
-        },
-      ],
-    },
-    {
-      name: 'BuilderBlockWithClassName',
-      component: BuilderBlockWithClassNameComponent,
-      shouldReceiveBuilderProps: {
-        builderBlock: true,
-        builderContext: true,
-        builderComponents: true,
-      },
-      inputs: [
-        {
-          name: 'content',
-          type: 'uiBlocks',
-          defaultValue: [
-            {
-              '@type': '@builder.io/sdk:Element',
-              '@version': 2,
-              id: 'builder-c6e179528dee4e62b337cf3f85d6496f',
-              component: {
-                name: 'Text',
-                options: {
-                  text: 'Enter some text...',
-                },
-              },
-              responsiveStyles: {
-                large: {
-                  display: 'flex',
-                  flexDirection: 'column',
-                  position: 'relative',
-                  flexShrink: '0',
-                  boxSizing: 'border-box',
-                  marginTop: '20px',
-                  lineHeight: 'normal',
-                  height: 'auto',
-                },
-              },
-            },
-          ],
-        },
-      ],
-    },
-  ];
+  customComponents: RegisteredComponent[] = customComponents;
 
   async ngOnInit() {
     const urlPath = window.location.pathname || '';
