+++ packages/sdks/e2e/angular/src/app/custom-components/builder-block-with-class-name.component.ts
