+++ packages/sdks/e2e/angular/src/app/custom-components/custom-button.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'button[custom-button]',
+  template: `
+    <span>
+      <ng-content></ng-content>
+    </span>
+  `,
+})
+export class CustomButtonComponent {}
