+++ packages/sdks/e2e/angular/src/app/custom-components/custom-card-content.component.ts
@@ -1,0 +1,18 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'div[custom-card-content]',
+  template: `
+    <div class="card-content">
+      <ng-content></ng-content>
+    </div>
+  `,
+  styles: [
+    `
+      .card-content {
+        padding: 8px 0;
+      }
+    `,
+  ],
+})
+export class CustomCardContentComponent {}
