+++ packages/sdks/e2e/angular/src/app/custom-components/custom-card.component.ts
@@ -1,0 +1,17 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'div[custom-card]',
+  template: ` <ng-content></ng-content> `,
+  styles: [
+    `
+      :host {
+        border: 1px solid #ccc;
+        border-radius: 4px;
+        padding: 16px;
+        margin-bottom: 16px;
+      }
+    `,
+  ],
+})
+export class CustomCardComponent {}
