+++ packages/sdks/e2e/angular/src/app/custom-components/hello.component.ts
